feat: reorganize project
