+++ src/components/Canvas/index.js
@@ -1,0 +1,3 @@
+// @ts-check
+
+export { CanvasBuilder } from './CanvasBuilder'
