+++ src/constants/color.js
@@ -1,6 +1,11 @@
 // @ts-check
 
-export const GRAVITY = { x: 0, y: 0.1 }
 export const RED = 'rgb(212, 42, 32)'
 export const YELLOW = 'rgb(250, 194, 43)'
 export const BLUE = 'rgb(14, 99, 142)'
+
+export const Color = Object.freeze({
+  RED,
+  YELLOW,
+  BLUE
+})
