+++ src/constants/physics.js
@@ -1,0 +1,7 @@
+// @ts-check
+
+export const GRAVITY = { x: 0, y: 0.1 }
+
+export const Physics = Object.freeze({
+  GRAVITY
+})
