+++ src/drawings/circle/index.js
@@ -1,0 +1,1 @@
+export { Circle } from './Circle'
