+++ src/drawings/example/Example.jsx
@@ -1,15 +1,14 @@
 // @ts-check
 
-import { CanvasBuilder } from './CanvasBuilder'
+import { CanvasBuilder } from '@/components/Canvas'
+import { Color } from '@/constants'
 
-/**
- * @type {import('./CanvasBuilder').DrawFactory}
- */
+/** @type {import('@/components/Canvas/CanvasBuilder').DrawFactory} */
 function drawExample (canvas, { scale }) {
   return {
     draw () {
       const ctx = canvas.getContext('2d')
-      ctx.fillStyle = 'green'
+      ctx.fillStyle = Color.YELLOW
       ctx.fillRect(
         0,
         0,
@@ -20,6 +19,6 @@ function drawExample (canvas, { scale }) {
   }
 }
 
-export const CanvasExample = new CanvasBuilder()
+export const Example = new CanvasBuilder()
   .withDrawFactory(drawExample)
   .build()
