+++ src/drawings/example/index.js
@@ -1,0 +1,1 @@
+export { Example } from './Example'
