+++ src/drawings/falling-ball/index.js
@@ -1,0 +1,1 @@
+export { FallingBall } from './FallingBall'
