--- src/drawings/index.html
@@ -1,54 +1,0 @@
-<!DOCTYPE html>
-
-<html lang="en">
-  <head>
-    <meta charset="utf-8">
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0, maximum-scale=1.0, user-scalable=0" />
-
-    <!--  STYLESHEETS  -->
-    <link href="https://fonts.googleapis.com/css?family=Inconsolata" rel="stylesheet">
-    <link rel="stylesheet" type="text/css" href="css/calculator.css"/>
-    <link rel="stylesheet" type="text/css" href="css/style.css"/>
-
-    <title>Calculator | PJM</title>
-  </head>
-  <body>
-
-    <header>
-      <a href="https://peterjmartinson.github.io"><div class="icon"></div></a>
-      <div class="name"><span>Canvas Testing Ground</span></div>
-    </header>
-
-    <div class="project">
-      <canvas id="canvas"></canvas>
-    </div>
-
-    <footer>
-      <div class="copyright">
-        Copyright 2023
-        <span>Peter J. Martinson</span>
-      </div>
-      <div class="contact">
-        Contact
-        <span>
-          <a href="mailto:peter.j.martinson@gmail.com">
-            peter.j.martinson@gmail.com
-          </a>
-        </span>
-      </div>
-      <div class="navigation">
-        Navigation
-        <span>
-          <a href="https://peterjmartinson.github.io/index.html">Home </a>&#151;
-          <a href="http://stonetelescope.wordpress.com" target="_blank">Stories </a>&#151;
-          <a href="https://twitter.com/StoneTelescope" target="_blank">Twitter</a>
-        </span>
-      </div>
-    </footer>
-
-    <script src="drawline.js"></script>
-
-  </body>
-</html>
-
